test_system.py
#!/usr/bin/env python3
"""
Test script to verify the complete AVEN AI system is working
"""

import sys
import json
import asyncio
from pathlib import Path

# Add the project root to Python path
project_root = Path(__file__).parent
sys.path.append(str(project_root))

async def test_system():
    """Test the complete system components"""
    print("🧪 Testing AVEN AI System Components")
    print("=" * 50)
    
    # Test 1: Import all required packages
    print("1. Testing package imports...")
    try:
        import google.generativeai as genai
        import pinecone
        from firecrawl import AsyncFirecrawlApp
        import numpy as np
        print("   ✅ All packages imported successfully")
    except ImportError as e:
        print(f"   ❌ Import error: {e}")
        return False
    
    # Test 2: Test environment variables
    print("\n2. Testing environment variables...")
    import os
    from dotenv import load_dotenv
    
    load_dotenv()
    
    required_vars = ['GEMINI_API_KEY', 'PINECONE_API_KEY', 'FIRECRAWL_API_KEY']
    missing_vars = []
    
    for var in required_vars:
        if not os.getenv(var):
            missing_vars.append(var)
    
    if missing_vars:
        print(f"   ❌ Missing environment variables: {missing_vars}")
        return False
    else:
        print("   ✅ All environment variables found")
    
    # Test 3: Test Pinecone connection
    print("\n3. Testing Pinecone connection...")
    try:
        from pinecone import Pinecone
        pc = Pinecone(api_key=os.getenv('PINECONE_API_KEY'))
        index = pc.Index('aven-v2')
        stats = index.describe_index_stats()
        print(f"   ✅ Connected to Pinecone. Vector count: {stats.total_vector_count}")
    except Exception as e:
        print(f"   ❌ Pinecone error: {e}")
        return False
    
    # Test 4: Test Gemini API
    print("\n4. Testing Gemini API...")
    try:
        genai.configure(api_key=os.getenv('GEMINI_API_KEY'))
        test_text = "Hello, this is a test"
        result = genai.embed_content(
            model="models/embedding-001",
            content=test_text,
            task_type="retrieval_document"
        )
        embedding = result['embedding']
        print(f"   ✅ Gemini API working. Embedding dimension: {len(embedding)}")
    except Exception as e:
        print(f"   ❌ Gemini API error: {e}")
        return False
    
    # Test 5: Test knowledge base query
    print("\n5. Testing knowledge base query...")
    try:
        import subprocess
        result = subprocess.run([
            sys.executable, 
            str(project_root / 'scripts' / 'query_knowledge_base.py'),
            'HELOC rates'
        ], capture_output=True, text=True, timeout=30)
        
        if result.returncode == 0:
            data = json.loads(result.stdout)
            if data.get('success') and data.get('results'):
                print(f"   ✅ Knowledge base query successful. Found {len(data['results'])} results")
            else:
                print("   ❌ Knowledge base query returned no results")
                return False
        else:
            print(f"   ❌ Knowledge base query failed: {result.stderr}")
            return False
    except Exception as e:
        print(f"   ❌ Knowledge base query error: {e}")
        return False
    
    # Test 6: Test Firecrawl (optional due to rate limits)
    print("\n6. Testing Firecrawl (optional)...")
    try:
        app = AsyncFirecrawlApp(api_key=os.getenv('FIRECRAWL_API_KEY'))
        print("   ✅ Firecrawl client initialized (not tested due to rate limits)")
    except Exception as e:
        print(f"   ⚠️  Firecrawl warning: {e}")
    
    print("\n" + "=" * 50)
    print("🎉 All system components are working correctly!")
    print("\nSystem Summary:")
    print("- Firecrawl: Web scraping with AsyncFirecrawlApp")
    print("- Gemini API: 768-dimensional embeddings")
    print("- Pinecone: Vector database with aven-v2 index")
    print("- Knowledge Base: Ready for queries")
    print("- Next.js API: Running on localhost:3001")
    print("- Vapi Integration: Configured with function calling")
    
    return True

if __name__ == "__main__":
    success = asyncio.run(test_system())
    sys.exit(0 if success else 1)
